+++ fluidSimLib/dataBuffer3.h
@@ -10,23 +10,23 @@ class dataBuffer3
 {
 private:
     size3 mSize;
-    std::vector<double> mData;
+    std::vector<FloatType> mData;
 
 public:
     dataBuffer3();
 
-    dataBuffer3( const size3& size, const double initValue = 0.0 );
+    dataBuffer3( const size3& size, const FloatType initValue = 0.0 );
 
-    dataBuffer3( size_t width, size_t height, size_t depth, const double initValue = 0.0 );
+    dataBuffer3( size_t width, size_t height, size_t depth, const FloatType initValue = 0.0 );
 
 public:
-    void resize( const size3& size, const double initValue = 0.0 );
+    void resize( const size3& size, const FloatType initValue = 0.0 );
 
-    void resize( size_t width, size_t height, size_t depth, const double initValue = 0.0 );
+    void resize( size_t width, size_t height, size_t depth, const FloatType initValue = 0.0 );
 
-    const double valueByIndex( size_t i, size_t j, size_t k ) const;
+    const FloatType valueByIndex( size_t i, size_t j, size_t k ) const;
 
-    double& valueByIndex( size_t i, size_t j, size_t k );
+    FloatType& valueByIndex( size_t i, size_t j, size_t k );
 
     template <typename Callback>
     void forEachIndex(Callback func) const
@@ -58,7 +58,7 @@ public:
         }
     }
 
-    void set( double value );
+    void set( FloatType value );
     void set( const dataBuffer3& value );
 
     size3 size( ) const;
@@ -69,20 +69,20 @@ public:
 
     size_t depth( ) const;
 
-    double *data( );
+    FloatType *data( );
 
-    std::vector<double>::iterator begin();
-    std::vector<double>::const_iterator begin() const;
-    std::vector<double>::iterator end();
-    std::vector<double>::const_iterator end() const;
+    std::vector<FloatType>::iterator begin();
+    std::vector<FloatType>::const_iterator begin() const;
+    std::vector<FloatType>::iterator end();
+    std::vector<FloatType>::const_iterator end() const;
 
     void swap( dataBuffer3& other );
 
-    double& operator()(size_t i, size_t j, size_t k);
-    const double& operator()( size_t i, size_t j, size_t k ) const;
+    FloatType& operator()(size_t i, size_t j, size_t k);
+    const FloatType& operator()( size_t i, size_t j, size_t k ) const;
     //
-    double& operator()( const size3& pt );
-    const double& operator()( const size3& pt ) const;
+    FloatType& operator()( const size3& pt );
+    const FloatType& operator()( const size3& pt ) const;
 };
 
 #endif
